Update README.md
Fixed the C-style example at the start to use the prefix-increment instead of postfix-increment. 

(Yes, both versions 'do the same' if your hope that the compiler will optimise the use of a postfix-increment for you. Using a prefix-increment is good style as it is more clear what the intention of the statement is (which is to increment the iteration index, and not to increment but return the old value). Nonetheless, many programmers (including myself for many years) do not know about this, and many tutorials/explanations list the postfix-version.

And this is of course exactly why iterator-based loops are so nice.